+++ backend-flask/services/create_activity.py
@@ -8,15 +8,18 @@ class CreateReply:
       'errors': None,
       'data': None
     }
+
     if cognito_user_id == None or len(cognito_user_id) < 1:
       model['errors'] = ['cognito_user_id_blank']
 
     if activity_uuid == None or len(activity_uuid) < 1:
       model['errors'] = ['activity_uuid_blank']
+
     if message == None or len(message) < 1:
       model['errors'] = ['message_blank'] 
     elif len(message) > 1024:
-      model['errors'] = ['message_exceed_max_chars_1024']
+      model['errors'] = ['message_exceed_max_chars'] 
+
     if model['errors']:
       # return what we provided
       model['data'] = {
