+++ bin/generate/migration
@@ -33,7 +33,7 @@ class {klass}Migration:
   def rollback():
     db.query_commit({klass}Migration.rollback_sql(),{{
     }})
-migration = AddBioColumnMigration
+migration = {klass}Migration
 """
 #remove leading and trailing new lines
 file_content = file_content.lstrip('\n').rstrip('\n')
