+++ frontend-react-js/src/components/ActivityActionReply.js
@@ -2,6 +2,7 @@ import {ReactComponent as ReplyIcon} from './svg/reply.svg';
 
 export default function ActivityActionReply(props) { 
   const onclick = (event) => {
+    console.log('activity-action-reply',props.activity)
     props.setReplyActivity(props.activity)
     props.setPopped(true)
   }
