+++ frontend-react-js/src/components/ActivityItem.css
@@ -3,6 +3,17 @@
   flex-direction: column;
   border-bottom: solid 1px rgb(60,54,79);
   overflow: hidden;
+}
+
+.replies {
+  padding-left: 24px;
+  background: rgba(255,255,255,0.15);
+}
+.replies.activity_item{
+  background: var(--fg);
+}
+
+.activity_main {
   padding: 16px;
 }
 .activity_item:last-child {
