+++ frontend-react-js/src/components/ActivityItem.js
@@ -24,12 +24,14 @@ export default function ActivityItem(props) {
 
   return (
     <div className='activity_item'>
-      <ActivityContent activity={props.activity} />
-      <div className="activity_actions">
-        <ActivityActionReply setReplyActivity={props.setReplyActivity} activity={props.activity} setPopped={props.setPopped} activity_uuid={props.activity.uuid} count={props.activity.replies_count}/>
-        <ActivityActionRepost activity_uuid={props.activity.uuid} count={props.activity.reposts_count}/>
-        <ActivityActionLike activity_uuid={props.activity.uuid} count={props.activity.likes_count}/>
-        <ActivityActionShare activity_uuid={props.activity.uuid} />
+      <div className="activity_main">
+        <ActivityContent activity={props.activity} />
+        <div className="activity_actions">
+          <ActivityActionReply setReplyActivity={props.setReplyActivity} activity={props.activity} setPopped={props.setPopped} activity_uuid={props.activity.uuid} count={props.activity.replies_count}/>
+          <ActivityActionRepost activity_uuid={props.activity.uuid} count={props.activity.reposts_count}/>
+          <ActivityActionLike activity_uuid={props.activity.uuid} count={props.activity.likes_count}/>
+          <ActivityActionShare activity_uuid={props.activity.uuid} />
+        </div>
       </div>
       {replies}
     </div>
